aider: Changed event listeners for markers to show info windows on hover and close them on mouseout.
# Aider chat conversation:

USER: awesome it worked. can you make the info windows appear just by hovering over a placemark

ASSISTANT: To implement the functionality where the info windows appear just by hovering over a placemark, we need to modify the `public\gmap.js` file. We will change the event listener for the markers from 'click' to 'mouseover' for showing the info window and add a 'mouseout' event to close the info window when the mouse leaves the marker.

Here are the *SEARCH/REPLACE* blocks for the necessary changes:

public\gmap.js
```javascript
<<<<<<< SEARCH
    buildingMarker.addListener('click', function () {
        buildingInfowindow.open(map, buildingMarker);
    });
=======
    buildingMarker.addListener('mouseover', function () {
        buildingInfowindow.open(map, buildingMarker);
    });
    buildingMarker.addListener('mouseout', function () {
        buildingInfowindow.close();
    });
>>>>>>> REPLACE
```

public\gmap.js
```javascript
<<<<<<< SEARCH
    marker.addListener('click', () => {
        infoWindow.open(map, marker);
    });
=======
    marker.addListener('mouseover', () => {
        infoWindow.open(map, marker);
    });
    marker.addListener('mouseout', () => {
        infoWindow.close();
    });
>>>>>>> REPLACE
```

These changes will make the info windows for both the tallest buildings and the comp markers appear when the user hovers over the placemarks and disappear when the mouse is moved away.